Rename Logbook.types.ts to LegacyLogbook.types.ts
