+++ frontend/src/features/Logbook/api.ts
@@ -5,7 +5,7 @@ import { monitorfishApi, monitorfishApiKy } from '../../api/api'
 import { HttpStatusCode } from '../../api/constants'
 import { ApiError } from '../../libs/ApiError'
 
-import type { VesselVoyage } from './Logbook.types'
+import type { VesselVoyage } from './LegacyLogbook.types'
 import type { VesselIdentity } from '../../domain/entities/vessel/types'
 
 const LOGBOOK_ERROR_MESSAGE = "Nous n'avons pas pu récupérer les messages JPE de ce navire"
