+++ frontend/src/features/Logbook/utils.ts
@@ -7,7 +7,7 @@ import { LayerProperties } from '../../domain/entities/layers/constants'
 import { undefinedize } from '../../utils/undefinedize'
 
 import type { CatchProperty, CatchWithProperties, ProtectedCatchWithProperties } from './components/VesselLogbook/types'
-import type { LogbookCatch, LogbookMessage, PNOMessageValue, ProtectedSpeciesCatch } from './Logbook.types'
+import type { LogbookCatch, LogbookMessage, PNOMessageValue, ProtectedSpeciesCatch } from './LegacyLogbook.types'
 import type { LogbookMessage as LogbookMessageNamespace } from './LogbookMessage.types'
 import type { SpeciesInsight, SpeciesToSpeciesInsight, SpeciesToSpeciesInsightList } from './types'
 import type { DeclaredLogbookSpecies, FishingActivityShowedOnMap } from '../../domain/entities/vessel/types'
